third_party/blink/web_tests/accessibility/listitem-presentation-inherited.html

<script src="../resources/js-test.js"></script>
<script src="../resources/accessibility-helper.js"></script>
<div id="container">
<ul role="presentation">
<li role="listitem">Item 1</li>
<li role="listitem">Item 2</li>
<li role="listitem">Item 3</li>
</ul>

<ul role="presentation">
<li>Plain Text 1</li>
<li>Plain Text 2</li>
<li>Plain Text 3</li>
</ul>
</div>  <!-- container -->
<p id="description"></p>
<div id="console"></div>
<script>
description("This tests that the presentation role is not inherited.");

if (window.testRunner)
    testRunner.dumpAsText();
if (window.accessibilityController) {
    buildAccessibilityTree(accessibilityController.accessibleElementById("container"), 0, 1);
}
</script>
