third_party/blink/web_tests/accessibility/listitem-presentation-inherited-expected.txt

Item 1
Item 2
Item 3
Plain Text 1
Plain Text 2
Plain Text 3

This tests that the presentation role is not inherited.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

AXRole: AXGenericContainer
    AXRole: AXListItem
        AXRole: AXNone "• "
            AXRole: AXStaticText "• "
                AXRole: AXInlineTextBox "• "
        AXRole: AXStaticText "Item 1"
            AXRole: AXInlineTextBox "Item 1"
    AXRole: AXListItem
        AXRole: AXNone "• "
            AXRole: AXStaticText "• "
                AXRole: AXInlineTextBox "• "
        AXRole: AXStaticText "Item 2"
            AXRole: AXInlineTextBox "Item 2"
    AXRole: AXListItem
        AXRole: AXNone "• "
            AXRole: AXStaticText "• "
                AXRole: AXInlineTextBox "• "
        AXRole: AXStaticText "Item 3"
            AXRole: AXInlineTextBox "Item 3"
    AXRole: AXNone
        AXRole: AXNone "• "
            AXRole: AXStaticText "• "
                AXRole: AXInlineTextBox "• "
        AXRole: AXStaticText "Plain Text 1"
            AXRole: AXInlineTextBox "Plain Text 1"
    AXRole: AXNone
        AXRole: AXNone "• "
            AXRole: AXStaticText "• "
                AXRole: AXInlineTextBox "• "
        AXRole: AXStaticText "Plain Text 2"
            AXRole: AXInlineTextBox "Plain Text 2"
    AXRole: AXNone
        AXRole: AXNone "• "
            AXRole: AXStaticText "• "
                AXRole: AXInlineTextBox "• "
        AXRole: AXStaticText "Plain Text 3"
            AXRole: AXInlineTextBox "Plain Text 3"

PASS successfullyParsed is true

TEST COMPLETE
